rename domain package to model
