+++ lib/model/highlighted_string.dart
