+++ lib/model/product.dart
