+++ lib/model/query.dart
