+++ lib/model/query_suggestion.dart
