+++ lib/ui/widgets/product_card_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_ecom_demo/domain/product.dart';
+import 'package:flutter_ecom_demo/model/product.dart';
 import 'package:flutter_ecom_demo/ui/app_theme.dart';
 import 'package:flutter_ecom_demo/ui/widgets/color_indicator_view.dart';
 import 'package:flutter_ecom_demo/ui/widgets/rating_view.dart';
